(maint) Use GUID in test service url
If the URLs we use are not unique, the tests tend to fail when run
consecutively, although there's nothing really wrong with the tests.

It seems the .Close() method takes a handful of seconds to actually
clean up the URL registration. Rather than mess about with arbitrary
sleep times, we can just use unique URLs.