(maint) Update nupkg packaging and build
Add back in acknowledgement for log4net, although it is only used
conditionally if the project depending on Rhino is already using it.

Update the csproj to include references to the license and
acknowledgements files, so that dotnet pack will put them in the nupkg.

Update the build script to build in release configuration by default.