(#7) Use SHA1 for license generation by default
This is consistent with the original .NET 4.0 implementation, and brings
the newer assembly versions back in line with the initial implementation
to ensure that there are no unexpected changes.

This also adds overloads for the Generate() method that allow consumers
to select a specific signing algorithm. This will allow us to opt-in to
newer license generation methods in future without recompiling this
assembly, should we want or need to.