(maint) Treat no warnings as errors
With Visual Studio 2026, if you do not explicitly set any warnings as
errors, it will use `$(WarningsAsErrors)` which appears to add `NU1605`
to the list. One of the referenced assemblies is referenced at version
1.0.2, and somehow this trips up the build by finding a 1.0.3 and trying
to use that? This appears to only be an issue building from within
Visual Studio, as the `recipe.cake` file handles this setting for us
during the build script execution.